Refactor scrollCursorIntoView
